pyproject.toml
[tool.poetry]
name = "mixs"
version = "6.3.0"
description = "MIxS: the Minimum Information about any (X) Sequence standard"
authors = [
    "Mark Andrew Miller <MAM@lbl.gov>",
    "Sujay Patil <spatil@lbl.gov>"
]
license = "CC0-1.0 license"
readme = "README.md"
include = ["README.md", "project", "src/mixs", "src/scripts"]

#[tool.poetry-dynamic-versioning]
#enable = true
#vcs = "git"
#style = "pep440"

[tool.poetry.dependencies]
python = "^3.10"
linkml-runtime = "^1.9"
jsonasobj2 = "^1.0" # used transitively in src/scripts/mixs/datamodel/mixs.py

[tool.poetry.group.dev.dependencies]
black = "^25.1"
deptry = "^0.23"
linkml = "^1.9"
matplotlib = "^3.10"
mkdocs-material = "^9.6"
mkdocs-mermaid2-plugin = "^1.2"
pandas = "^2.3"
pyarrow = "^21.0"
rdflib = "^7.1"
schemasheets = "^0.4"
scipy = "^1.15"
yamllint = "^1.37"
click = "^8.1"
pyyaml = "^6.0"
requests = "^2.31"
python-dotenv = "^1.0"

# Optional dependency group for legacy-diff tool
# These dependencies are only needed for comparing historical MIxS schemas
# (Excel, Word, XSD formats from v1-v5). Install with: poetry install --with legacy-diff
[tool.poetry.group.legacy-diff]
optional = true

[tool.poetry.group.legacy-diff.dependencies]
openpyxl = "^3.1.0"      # .xlsx reading (MIxS v5, 2018)
xlrd = ">=1.2.0,<2.0.0"  # .xls reading (MIxS v2-v4, 2009-2014) - xlrd 2.0+ dropped .xls support
python-docx = "^1.0.0"   # .docx reading (pre-2009 Word documents)

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry.extras]
docs = ["linkml", "mkdocs-material"]

[tool.poetry.scripts]
extension-distances = 'scripts.extension_distances:generate_dendrogram'
extension-differences = 'scripts.extension_slot_differences:set_arithmetic'
linkml2class-tsvs = 'scripts.linkml2class_tsvs:process_schema_classes'
generate-combinations = 'scripts.combinations_list_generator:generate_combinations'
generate-enumerations = 'scripts.enumerations_list_generator:generate_enumerations'
generate-term-list = 'scripts.term_list_generator:generate_term_list'
organize-files = 'scripts.organize_files:main'
required-supersedes-recommended = 'scripts.required_supersedes_recommended:find_required_and_recommended_usages'
describe-enums = 'scripts.describe_enums_by_slots_using:update_enum_descriptions'
isolate-global-slots = 'scripts.isolate_slots:isolate_global_slots'
diff-releases = 'scripts.diff_two_linkml_mixs_releases:main'
mixs-legacy-diff = 'mixs.diff.cli:main'

[tool.deptry.package_module_name_map]
pyyaml = "yaml"
python-dotenv = "dotenv"
python-docx = "docx"
